HT-1459: Add active, expired scopes
- We could use placeholders with Time.now, but sqlite3 and mysql both
support CURRENT_TIMESTAMP, so this seems less brittle.

- Requires email, userid, expires fields: this would be enforced by
the database in production. Missing email in development/test leads to
odd behavior -- users with NULL email are not equal to themselves. This
functionality was essentially required to get the tests for
expires/active to pass, so I did not add separate tests for it.