Build an actual object for the expiration_date
The expiration date is being manipulated enough that
it's worth breaking out into its own class that
can compute things (e.g., whether it's expired,
how long until it expires, etc.)

Expiration dates are always dates only; the
time is thrown away. A quick question to Melissa
verified that this is ok.
https://hathitrust.slack.com/archives/DKV93G37T/p1576770049002400

The HTUser object now creates a new expiration date
object on creation and whenever `expires` is set and
forwards appropriate methods to it. The
expriation date object itself is read-only.