HT-2007: ensure iprestrict remains nil on update
The behavior if iprestrict is an empty string (which happens by default
when editing a user) on the back end is undefined, so it's preferable to
ensure it stays nil. This also breaks the validation into two parts, one
to ensure there is an iprestriction unless MFA is true, and the other to
validate the format.