HT-2390: hash one-time link instead of encrypting
The application doesn't have a need for recoverable tokens, so instead
of encrypting using the application secret, just use the SHA256 hash of
the token.  There isn't really a need for salting here since the
password is completely random and long enough that a rainbow table
attack wouldn't be feasible (and in the end we're sending these in
emails anyway).

If we feel we need more security we could use bcrypt, but we would also
need to send the id of the approval request so that we can find the
relevant approval request without examining each one.